only store node data in Node objects
rather than both in Node objects and in the dict of G.nodes.

This means we don't need to save() or update() the node objects
as there is now only one source of truth.  Under the hood the dict of
G.nodes has a single key "node_obj" which has the value of the Node
object.